server/requirements.txt
Django==1.11.18
django-activeurl==0.1.9
django-appconf==1.0.2
django-bootstrap3==9.0.0
django-bootstrap4==0.0.1
djangorestframework==3.6.3
ipaddress==1.0.18
virtualenv==1.11.6
mysql-python==1.2.5
sqlparse==0.2.4